Working method 1 for Get next line project
This algorithm uses string join function and appends every read call until new line character is found in the static variable. Memory in heap is allocated for the number of characters until new line or end of string from the static variable. And static variable is updated and ready for further calls of the functions.